mqttnxd: rework
This commit reworks mqttnxd.
EIB is not forwarded as is into MQTT.
EIB is mirrored/synchronised to MQTT.
The MQTT topic names are not chosen to indicate EIB addresses.
mqttnxd is configured by (retained) MQTT topics with a specific suffix.
The corresponding MQTT topic will then be mirrored to the
configured EIB address.
The MQTT end does not have a specific 'set' topic.
The main topic is mirrored to EIB (and vice versa).